requirements.txt
torch
torchvision
pillow
matplotlib
piq
lpips
